job name and 'source' command
a) Added the ability to change the job name on cluster through the 'add' function argument.

b) Added the ability to insert the 'source' command into the job startup file, which, by adding the line 'source /etc/profile.d/modules.sh' in simanrc.py, resolves the issue of correct automatic startup on the cluster.